test(component-library): remove assertion from clipboard test
Testing the clipboard is not very easy. It appears to be
flaky in storybook and it's not easy to do with unit tests
a solution would be to use vitest's browser mode and use
an actual browser to test the feature.

That's how @vue/test-utils does it (the utility we use
to create the copy feature), they have separate vitest browser
test to make sure the copying works.